backend/requirements.txt
beautifulsoup4
lxml
mysql-connector-python
requests
